Ensure all bots are returned during getProductIotMap
During the use of the Ecovacs Home App, some bots do disappear from the
app. This is due to the getProductIotMap endpoint handler using its own
list of bots. This list contains only a fraction of the
EcoVacsHomeProducts list maintained in models.py. Replacing the handler
list with the global one follows DRY and simplifies maintenance in the
future.